+++ Code/main_impt.py
@@ -1,0 +1,3 @@
+from systems.server_system import ServerSystem
+
+server_system = ServerSystem()
