+++ Code/systems/bin_system/__init__.py
@@ -1,0 +1,1 @@
+from systems.bin_system.bin_system import BinaryFileSystem
