+++ Code/systems/texture_system/texture_system.py
