--- Code/texture_system/__init__.py
@@ -1,1 +1,0 @@
-from texture_system.texture_system import TextureSystem
