--- console client/server_system_setup.py
@@ -1,3 +1,0 @@
-from server_system import ServerSystem
-
-server_system = ServerSystem()
